+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.2.9 - 2020-10-27
+
+### Changed
+
+* Removed daily and weekly subscriptions lengths
+
+### Fixed
+
+* Fixed error where lineitem didn't take subscription length into consideration when calculating pric
+
 ## 2.2.8 - 2020-10-26
 
-## Added
+### Added
 
 * Added `couponCode` to the payment order